Implementing the method to delete a password and allowing DELETE request in the cors middlewhere
